move the `ByteTypes` definition from `utils.py` to `odxtypes.py`
this is more consistent with the other typedefs like `AtomicOdxType`,
`ParameterValue`, etc.
